Add babyoil hazard, respawn, and checkpoints
Created a killzone boundary that follows the camera on the bottom of the screen. Need to add texture to it to show its baby oil. Added checkpoint prefab that is reusable as a respawn point. Added respawn script that teleports the player back to the last checkpoint or inital spawn point.