Add texture to babyoil hazard, dynamically adjust boundaries to screen size
Added a texture for the babyoil hazard aswell as a script that moves the texture with the camera. Added dynamically adjusting left and top boundaries that change based on resolution size. Added script to adjust babyoil percentage of screen coverage and dynamically adjust killzone box based on height of baby oil.